맵오브젝트, 그리고 관련 메소드,변수들이 묶여있는 클래스들 변경...대충 뭔지 정확히 짐작은 안가나 일단 타일에 찡기고, 날아디는 현상은 고침

+++ src/com/teamcriminals/TileMap/TileMap.java
@@ -112,12 +112,12 @@ public class TileMap {
 		return tileSize;
 	}
 
-	public double getx() {
-		return x;
+	public int getx() {
+		return (int) x;
 	}
 
-	public double gety() {
-		return y;
+	public int gety() {
+		return (int) y;
 	}
 
 	public int getWidth() {
